New clientlib ver

+++ Kaltura/Client/Type/StatsEvent.php
@@ -50,7 +50,7 @@ class Kaltura_Client_Type_StatsEvent extends Kaltura_Client_ObjectBase
 		if(count($xml->eventType))
 			$this->eventType = (int)$xml->eventType;
 		if(count($xml->eventTimestamp))
-			$this->eventTimestamp = (int)$xml->eventTimestamp;
+			$this->eventTimestamp = (float)$xml->eventTimestamp;
 		if(count($xml->sessionId))
 			$this->sessionId = (string)$xml->sessionId;
 		if(count($xml->partnerId))
@@ -117,7 +117,7 @@ class Kaltura_Client_Type_StatsEvent extends Kaltura_Client_ObjectBase
 	/**
 	 * the client's timestamp of this event
 	 *
-	 * @var int
+	 * @var float
 	 */
 	public $eventTimestamp = null;
